Added other pages layout cleanup.
