+++ CoreLocalizationTest/Controllers/ServicesController.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace CoreLocalizationTest.Controllers
+{
+    public class ServicesController : BaseController
+    {
+        // GET
+        [HttpGet("/{culture:regex(en)}/dashboard/services")]
+        [HttpGet("/{culture:regex(fr)}/tableaudebord/services")]
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
