+++ CoreLocalizationTest/Controllers/TestController.cs
@@ -1,12 +1,8 @@
-﻿using System.Linq;
-using System.Text.RegularExpressions;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Controllers;
-using Microsoft.AspNetCore.Mvc.Filters;
+﻿using Microsoft.AspNetCore.Mvc;
 
 namespace CoreLocalizationTest.Controllers
 {
-    public class TestController : Controller
+    public class TestController : BaseController
     {
         [HttpGet("/{culture:regex(en)}/test/about")]
         [HttpGet("/{culture:regex(fr)}/tester/apropos")]
@@ -18,30 +14,6 @@ namespace CoreLocalizationTest.Controllers
         public IActionResult Index()
         {
             return View();
-        }
-
-        public override void OnActionExecuting(ActionExecutingContext context)
-        {
-            var culture = context.RouteData.Values["culture"].ToString();
-            var alternateCulture = culture.Equals("en") ? "fr" : "en";
-            var path = string.Empty;
-
-            if (context.ActionDescriptor is ControllerActionDescriptor controllerActionDescriptor)
-            {
-                var actionAttributes = controllerActionDescriptor.MethodInfo.GetCustomAttributes(inherit: true)
-                    .Cast<HttpGetAttribute>()
-                    .SingleOrDefault(x => !x.Template.Contains(culture));
-
-                var regex = new Regex(@"/{(.*?)}");
-                if (actionAttributes != null)
-                {
-                    path = regex.Replace(actionAttributes.Template, "");
-                }
-            }
-
-            ViewData["Toggle"] = $"/{alternateCulture}{path}";
-
-            base.OnActionExecuting(context);
         }
     }
 }
