+++ CoreLocalizationTest/Resources/SharedResource.fr.resx
@@ -120,7 +120,16 @@
   <data name="Hello" xml:space="preserve">
     <value>Hello en francais!</value>
   </data>
+  <data name="Home" xml:space="preserve">
+    <value>Page d'accueil</value>
+  </data>
   <data name="Language" xml:space="preserve">
     <value>English</value>
+  </data>
+  <data name="ServiceRequests" xml:space="preserve">
+    <value>Mes requêtes</value>
+  </data>
+  <data name="Services" xml:space="preserve">
+    <value>Mes services</value>
   </data>
 </root>
