+++ CoreLocalizationTest/Views/ServiceRequests/Index.cshtml
@@ -1,0 +1,3 @@
+﻿<div>
+    <h2>@Localizer["ServiceRequests"]</h2>
+</div>
