+++ CoreLocalizationTest/Views/Services/Index.cshtml
@@ -1,0 +1,3 @@
+﻿<div>
+    <h2>@Localizer["Services"]</h2>
+</div>
