+++ CoreLocalizationTest/Views/Test/About.cshtml
@@ -1,4 +1,4 @@
 ﻿
     <div>
-        <h1>@Localizer["Hello"]</h1>
+        <h2>@Localizer["Hello"]</h2>
     </div>
